Add files via upload
This is the GUI for the project. It includes a JTree to list the websites in the first level of the tree, and usernames beyond that. There are buttons to add a website, add a username, and delete items. The delete button will delete a website along with any usernames. As of the upload, it is missing the ability to work with all other classes. It just holds placeholders for the set password, get password, and generate random password buttons.